requirements.txt
# Dependencies
fastapi
uvicorn
python-multipart
google-generativeai
deepgram-sdk
elevenlabs
langchain
langchain-community
langchain-google-genai
faiss-cpu
streamlit
python-dotenv
requests
numpy
